fix: Address review feedback for massage entities
- keeson.py: Set supports_massage_timer to False since we can't
  reliably emulate stepping without knowing current state
- number.py: Simplify level calculation to just round(value)
- select.py: Validate option against allowed options before processing,
  normalize timer_mode to return "Off" for invalid/unknown values
- sensor.py: Normalize timer_mode in native_value - treat "0", 0, or
  empty string as "Off"

+++ custom_components/adjustable_bed/number.py
@@ -393,7 +393,7 @@ class AdjustableBedMassageNumber(AdjustableBedEntity, NumberEntity):
     async def async_set_native_value(self, value: float) -> None:
         """Set the massage intensity level."""
         zone = self.entity_description.massage_zone
-        level = int(round(value))
+        level = round(value)
 
         _LOGGER.info(
             "Massage intensity set requested: %s zone to level %d (device: %s)",
